PDI-2365 : Repository : simplify, abstract, turn into an interface
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@10631 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/org/pentaho/di/trans/steps/mapping/MappingIODefinition.java
@@ -129,7 +129,7 @@ public class MappingIODefinition implements Cloneable {
 		return xml.toString();
 	}
 
-	public void saveRep(Repository rep, long id_transformation, long id_step, String prefix, long nr) throws KettleException {
+	public void saveRep(Repository rep, long id_transformation, long id_step, String prefix, int nr) throws KettleException {
 		rep.saveStepAttribute(id_transformation, id_step, nr, prefix+"input_step", inputStepname); 
 		rep.saveStepAttribute(id_transformation, id_step, nr, prefix+"output_step", outputStepname); 
 		rep.saveStepAttribute(id_transformation, id_step, nr, prefix+"main_path", mainDataPath);
